Unified run.py file allowing execution of an excavator against a file.
Provide a central excavators module which re-exports the various named
configurations. Use this to implement a CLI recognizing a --excavator
argument which will call an exacavtion of the supplied filename.

Rework run_example as a wrapper around this new unified CLI. Expand the
README documenting both run_example and unified run entry points noting
that in the absence of arguments a list of supported excavators is output.

While here add smoke tests for basic validation of exports and scripts.