Create basic tests covering excavation & output generation for example.
Make the minimal set of changes such that we are able to trigger the
excavation to a known directory of the in-tree example file. Check
that the expected HTML files were written for this known excavation
and assert some basic properties of the excavated arifacts.

The commit includes reworking of run_example wherein its excavation is
declared as a class and its core logic separated out into a run file.
This is done both to allow the tests to easily make use of the logic
and as groundwork for subsequent work to provide a unified entry point.